Save merged functioning login page
