+++ frontend/src/AuthContext.js
@@ -6,6 +6,9 @@ import jwtDecode from 'jwt-decode';
 
 export const AuthContext = createContext(null);
 
+const API_BASE = process.env.REACT_APP_API_BASE;
+
+
 export function useAuth() {
   return useContext(AuthContext);
 }
